+++ lending/src/sdk/models/operations/listaccountingbankaccounttransactions.ts
@@ -3,7 +3,7 @@
  */
 
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
-import * as shared from "../shared";
+import * as shared from "../../../sdk/models/shared";
 import { AxiosResponse } from "axios";
 
 export class ListAccountingBankAccountTransactionsRequest extends SpeakeasyBase {
@@ -79,5 +79,5 @@ export class ListAccountingBankAccountTransactionsResponse extends SpeakeasyBase
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }