+++ lending/src/sdk/models/operations/listaccountingdirectincomeattachments.ts
@@ -3,7 +3,7 @@
  */
 
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
-import * as shared from "../shared";
+import * as shared from "../../../sdk/models/shared";
 import { AxiosResponse } from "axios";
 
 export class ListAccountingDirectIncomeAttachmentsRequest extends SpeakeasyBase {
@@ -55,5 +55,5 @@ export class ListAccountingDirectIncomeAttachmentsResponse extends SpeakeasyBase
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }