+++ lending/src/sdk/models/operations/listcommercelocations.ts
@@ -3,7 +3,7 @@
  */
 
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
-import * as shared from "../shared";
+import * as shared from "../../../sdk/models/shared";
 import { AxiosResponse } from "axios";
 
 export class ListCommerceLocationsRequest extends SpeakeasyBase {
@@ -34,7 +34,7 @@ export class ListCommerceLocationsResponse extends SpeakeasyBase {
     contentType: string;
 
     /**
-     * Your `query` parameter was not correctly formed
+     * Your API request was not properly authorized.
      */
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;
@@ -49,5 +49,5 @@ export class ListCommerceLocationsResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }