+++ lending/src/sdk/models/operations/listdataintegritydetails.ts
@@ -3,7 +3,7 @@
  */
 
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
-import * as shared from "../shared";
+import * as shared from "../../../sdk/models/shared";
 import { AxiosResponse } from "axios";
 
 export class ListDataIntegrityDetailsRequest extends SpeakeasyBase {
@@ -58,7 +58,7 @@ export class ListDataIntegrityDetailsResponse extends SpeakeasyBase {
     dataIntegrityDetails?: shared.DataIntegrityDetails;
 
     /**
-     * Your API request was not properly authorized.
+     * Your `query` parameter was not correctly formed
      */
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;
@@ -73,5 +73,5 @@ export class ListDataIntegrityDetailsResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }