+++ lending/src/sdk/models/shared/pulloperation.ts
@@ -3,7 +3,6 @@
  */
 
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
-import { DataType } from "./datatype";
 import { Expose } from "class-transformer";
 
 /**
@@ -84,11 +83,11 @@ export class PullOperation extends SpeakeasyBase {
     connectionId: string;
 
     /**
-     * Available Data types
+     * The data type you are requesting in a pull operation.
      */
     @SpeakeasyMetadata()
     @Expose({ name: "dataType" })
-    dataType: DataType;
+    dataType: string;
 
     /**
      * A message about a transient or persistent error.
@@ -105,7 +104,7 @@ export class PullOperation extends SpeakeasyBase {
     id: string;
 
     /**
-     * `True` if the pull operation completed successfully.
+     * `True` if the pull operation is completed successfully. The `isCompleted` property is not queryable. To filter failed pull operations, query by `status!=Complete&&status!=NotSupported` instead.
      */
     @SpeakeasyMetadata()
     @Expose({ name: "isCompleted" })
@@ -158,4 +157,11 @@ export class PullOperation extends SpeakeasyBase {
     @SpeakeasyMetadata()
     @Expose({ name: "status" })
     status: PullOperationStatus;
+
+    /**
+     * Additional information about the dataset status.
+     */
+    @SpeakeasyMetadata()
+    @Expose({ name: "statusDescription" })
+    statusDescription?: string;
 }